+++ lib/synchronousdeluge/exceptions.py
@@ -1,5 +1,6 @@
 __all__ = ["DelugeRPCError"]
 
+
 class DelugeRPCError(Exception):
     def __init__(self, name, msg, traceback):
         self.name = name
@@ -8,4 +9,3 @@ class DelugeRPCError(Exception):
 
     def __str__(self):
         return "{0}: {1}: {2}".format(self.__class__.__name__, self.name, self.msg)
-