+++ medusa/server/web/config/providers.py
@@ -365,11 +365,11 @@ class ConfigProviders(Config):
                 provider.enable_manualsearch = 0  # these exceptions are actually catching unselected checkboxes
 
         if hasattr(provider, 'enable_search_delay'):
-                try:
-                    provider.enable_search_delay = config.checkbox_to_value(
-                        kwargs['{id}_enable_search_delay'.format(id=provider.get_id())])
-                except (AttributeError, KeyError):
-                    provider.enable_search_delay = 0  # these exceptions are actually catching unselected checkboxes
+            try:
+                provider.enable_search_delay = config.checkbox_to_value(
+                    kwargs['{id}_enable_search_delay'.format(id=provider.get_id())])
+            except (AttributeError, KeyError):
+                provider.enable_search_delay = 0  # these exceptions are actually catching unselected checkboxes
 
         if hasattr(provider, 'search_delay'):
             try: