+++ medusa/show/recommendations/anidb.py
@@ -13,7 +13,10 @@ from medusa.indexers.indexer_config import INDEXER_TVDBV2
 from medusa.logger.adapters.style import BraceAdapter
 from medusa.session.core import MedusaSession
 from medusa.show.recommendations.recommended import (
-    MissingTvdbMapping, RecommendedShow, cached_aid_to_tvdb, create_key_from_series
+    MissingTvdbMapping,
+    RecommendedShow,
+    cached_aid_to_tvdb,
+    create_key_from_series,
 )
 
 from simpleanidb import Anidb, REQUEST_HOT
@@ -37,7 +40,7 @@ class AnidbPopular(object):  # pylint: disable=too-few-public-methods
         self.default_img_src = 'poster.png'
 
     @recommended_series_cache.cache_on_arguments(namespace='anidb', function_key_generator=create_key_from_series)
-    def _create_recommended_show(self, series, storage_key=None):
+    def _create_recommended_show(self, storage_key, series):
         """Create the RecommendedShow object from the returned showobj."""
         try:
             tvdb_id = cached_aid_to_tvdb(series.aid)
@@ -85,7 +88,8 @@ class AnidbPopular(object):  # pylint: disable=too-few-public-methods
 
         for show in series:
             try:
-                recommended_show = self._create_recommended_show(show, storage_key='anidb_{0}'.format(show.aid))
+                recommended_show = self._create_recommended_show(storage_key=show.aid,
+                                                                 series=show)
                 if recommended_show:
                     result.append(recommended_show)
             except MissingTvdbMapping: