+++ medusa/tv/episode.py
@@ -304,6 +304,12 @@ class Episode(TV):
                 )
                 return super(Episode, self).__getattribute__(refactor)
 
+    def __eq__(self, other):
+        """Override default equalize implementation."""
+        return all([self.series.identifier == other.series.identifier,
+                    self.season == other.season,
+                    self.episode == other.episode])
+
     @classmethod
     def find_by_series_and_episode(cls, series, episode_number):
         """Find Episode based on series and episode number.
@@ -647,6 +653,7 @@ class Episode(TV):
             self.indexerid = int(sql_results[0]['indexerid'])
             self.indexer = int(sql_results[0]['indexer'])
 
+            # FIXME: This shouldn't be part of a possible apiv2 episodes request
             xem_refresh(self.series)
 
             self.scene_season = try_int(sql_results[0]['scene_season'], 0)
@@ -868,7 +875,7 @@ class Episode(TV):
                 '{id}: {series} episode statuses unchanged. Location is missing: {location}', {
                     'id': self.series.series_id,
                     'series': self.series.name,
-                    'location': self.series.raw_location,
+                    'location': self.series.location,
                 }
             )
             return
@@ -1273,6 +1280,7 @@ class Episode(TV):
                          self.is_proper, self.series.series_id, self.season, self.episode, self.absolute_number,
                          self.version, self.release_group, self.manually_searched, self.watched, self.preferred_words_score, ep_id]]
             else:
+                # @TODO: I checked this query, but it's not up2date with the db. We should remove it or review it.
                 # use a custom insert method to get the data into the DB.
                 return [
                     'INSERT OR IGNORE INTO '
@@ -2067,9 +2075,3 @@ class Episode(TV):
                     'filepath': filepath,
                 }
             )
-
-    def __eq__(self, other):
-        """Override default equalize implementation."""
-        return all([self.series.identifier == other.series.identifier,
-                    self.season == other.season,
-                    self.episode == other.episode])