+++ themes-default/slim/src/components/login.vue
@@ -19,5 +19,5 @@ export default {
 </script>
 
 <style>
-/* placeholder */
+
 </style>