+++ themes-default/slim/src/store/modules/defaults.js
@@ -1,19 +1,32 @@
 const state = {
     show: {
         airs: null,
+        airsFormatValid: null,
         akas: null,
         cache: null,
         classification: null,
         config: {
             airByDate: null,
-            aliases: null,
+            aliases: [],
             anime: null,
             defaultEpisodeStatus: null,
             dvdOrder: null,
             location: null,
+            locationValid: null,
             paused: null,
-            qualities: null,
-            release: null,
+            qualities: {
+                allowed: [],
+                preferred: []
+            },
+            release: {
+                requiredWords: [],
+                ignoredWords: [],
+                blacklist: [],
+                whitelist: [],
+                allgroups: [],
+                requiredWordsExclude: null,
+                ignoredWordsExclude: null
+            },
             scene: null,
             seasonFolders: null,
             sports: null,
@@ -21,13 +34,29 @@ const state = {
             airdateOffset: null
         },
         countries: null,
-        country_codes: null, // eslint-disable-line camelcase
-        genres: null,
+        genres: [],
         id: {
             tvdb: null,
             slug: null
         },
         indexer: null,
+        imdbInfo: {
+            akas: null,
+            certificates: null,
+            countries: null,
+            countryCodes: null,
+            genres: null,
+            imdbId: null,
+            imdbInfoId: null,
+            indexer: null,
+            indexerId: null,
+            lastUpdate: null,
+            plot: null,
+            rating: null,
+            runtimes: null,
+            title: null,
+            votes: null
+        },
         language: null,
         network: null,
         nextAirDate: null,
@@ -43,7 +72,10 @@ const state = {
         status: null,
         title: null,
         type: null,
-        year: {}
+        year: {},
+        size: null,
+        showQueueStatus: [],
+        xemNumbering: []
     }
 };