+++ themes-default/slim/src/store/modules/shows.js
@@ -3,7 +3,11 @@ import { api } from '../../api';
 import { ADD_SHOW } from '../mutation-types';
 
 const state = {
-    shows: []
+    shows: [],
+    currentShow: {
+        indexer: null,
+        id: null
+    }
 };
 
 const mutations = {
@@ -28,6 +32,10 @@ const mutations = {
         // Update state
         Vue.set(state.shows, state.shows.indexOf(existingShow), newShow);
         console.debug(`Merged ${newShow.title || newShow.indexer + String(newShow.id)}`, newShow);
+    },
+    currentShow(state, { indexer, id }) {
+        state.currentShow.indexer = indexer;
+        state.currentShow.id = id;
     }
 };
 
@@ -41,6 +49,9 @@ const getters = {
     getEpisode: state => ({ id, indexer, season, episode }) => {
         const show = state.shows.find(show => Number(show.id[indexer]) === Number(id));
         return show && show.seasons && show.seasons[season] ? show.seasons[season][episode] : undefined;
+    },
+    getCurrentShow: (state, getters, rootState) => {
+        return state.shows.find(show => Number(show.id[state.currentShow.indexer]) === Number(state.currentShow.id)) || rootState.defaults.show;
     }
 };
 
@@ -62,16 +73,26 @@ const actions = {
      * @returns {Promise} The API response.
      */
     getShow(context, { indexer, id, detailed, fetch }) {
-        const { commit } = context;
-        const params = {};
-        if (detailed !== undefined) {
-            params.detailed = Boolean(detailed);
-        }
-        if (fetch !== undefined) {
-            params.fetch = Boolean(fetch);
-        }
-        return api.get('/series/' + indexer + id, { params }).then(res => {
-            commit(ADD_SHOW, res.data);
+        return new Promise((resolve, reject) => {
+            const { commit } = context;
+            const params = {};
+
+            if (detailed !== undefined) {
+                params.detailed = Boolean(detailed);
+            }
+
+            if (fetch !== undefined) {
+                params.fetch = Boolean(fetch);
+            }
+
+            api.get('/series/' + indexer + id, { params })
+                .then(res => {
+                    commit(ADD_SHOW, res.data);
+                    resolve(res.data);
+                })
+                .catch(error => {
+                    reject(error);
+                });
         });
     },
     /**
@@ -84,17 +105,42 @@ const actions = {
     getShows(context, shows) {
         const { commit, dispatch } = context;
 
-        // If no shows are provided get the first 10k
+        // If no shows are provided get the first 1000
         if (!shows) {
-            const params = {
-                limit: 10000
-            };
-            return api.get('/series', { params }).then(res => {
-                const shows = res.data;
-                return shows.forEach(show => {
-                    commit(ADD_SHOW, show);
-                });
-            });
+            return (() => {
+                const limit = 1000;
+                const page = 1;
+                const params = {
+                    limit,
+                    page
+                };
+
+                // Get first page
+                api.get('/series', { params })
+                    .then(response => {
+                        const totalPages = Number(response.headers['x-pagination-total']);
+                        response.data.forEach(show => {
+                            commit(ADD_SHOW, show);
+                        });
+
+                        // Optionally get additional pages
+                        const pageRequests = [];
+                        for (let page = 2; page <= totalPages; page++) {
+                            const newPage = { page };
+                            newPage.limit = params.limit;
+                            pageRequests.push(api.get('/series', { params: newPage }).then(response => {
+                                response.data.forEach(show => {
+                                    commit(ADD_SHOW, show);
+                                });
+                            }));
+                        }
+
+                        return Promise.all(pageRequests);
+                    })
+                    .catch(() => {
+                        console.log('Could not retrieve a list of shows');
+                    });
+            })();
         }
 
         return shows.forEach(show => dispatch('getShow', show));